Moved hard-coded /tmp TMPDIR to Dockerfile.
Previously, use of /tmp as the temporary directory
was hard-coded into every data pipeline tool run.

When running Alphafold outside of Docker (e.g. on a cluster where
/tmp may be a size-limited directory and the environment variable
TMPDIR is set), this hard-coding may cause the /tmp directory to
run out of space.

By removing these hard-coded paths and relying on the mkdtemp
behavior (which checks the environment variables TMPDIR among others)
and setting TMPDIR in the Dockerfile, we can make this easier to run
both inside and outside of Docker.