Add FAQ explaining that colabs can time out
It can be confusing to read errors indicating missing python libraries if the imported library is cleared from memory! Since the cells in the notebook can take several hours to run, it is quite likely that a user of the colab would leave a cell running for a while and then come back later to run the next one. If the first cell has then timed out it can be difficult to debug what's going on if you don't know that colabs can time out.