Do not upgrade jax in Dockerfile
The Dockerfile pins jaxlib and then upgrades jax to the newest version.  This leads to a jax version being installed that is incompatible with the pinned jaxlib version.  See #298 

So everyone is working with the same versions, jax should not be upgraded.   This change accomplishes this.

Alternatively, if DeepMind validates the newest jax/jaxlib versions give correct results, then newer versions of jax/jaxlib should be pinned (but this same Dockerfile change should be used).

Closes #298