Add option to use precomputed templates
My experience is that template processing takes a huge fraction of the
overall runtime. It's also all on CPU, which is a waste if you're using
GPU machines.

Note that this uses the features.pkl which is already saved by the
existing code. For more stability across numpy versions, etc. we could
instead use `npz` files with `numpy.save`/`numpy.load` as `features.pkl`
just holds a dictionary of numpy arrays. I wanted to first gauge the
reaction to this general idea though.

See discussion in
https://github.com/google-deepmind/alphafold/issues/895